fix: update indirect call executor and registry to use Identity instead of AccountId
