+++ bitacross-worker/app-libs/parentchain-interface/Cargo.toml
@@ -28,10 +28,12 @@ sp-core = { default-features = false, features = ["full_crypto"], git = "https:/
 sp-runtime = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
 
 # litentry
+bc-relayer-registry = { path = "../../bitacross/core/bc-relayer-registry", default-features = false }
 lc-scheduled-enclave = { path = "../../litentry/core/scheduled-enclave", default-features = false, optional = true }
 litentry-hex-utils = { path = "../../../primitives/hex", default-features = false }
 litentry-primitives = { path = "../../litentry/primitives", default-features = false }
 sp-std = { default-features = false, git = "https://github.com/paritytech/substrate.git", branch = "polkadot-v0.9.42" }
+
 
 [dev-dependencies]
 env_logger = "0.9.0"
@@ -66,6 +68,7 @@ std = [
     "litentry-primitives/std",
     "lc-scheduled-enclave/std",
     "sp-std/std",
+    "bc-relayer-registry/std",
 ]
 sgx = [
     "sgx_tstd",
@@ -77,4 +80,5 @@ sgx = [
     "itp-top-pool-author/sgx",
     "litentry-primitives/sgx",
     "lc-scheduled-enclave/sgx",
+    "bc-relayer-registry/sgx",
 ]
