+++ bitacross-worker/bitacross/core/bc-relayer-registry/Cargo.toml
@@ -1,0 +1,83 @@
+[package]
+name = "bc-relayer-registry"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+bitcoin = { version = "0.31.0", default-features = false, features = ["secp-recovery", "no-std"] }
+codec = { package = "parity-scale-codec", version = "3.0.0", default-features = false, features = ["derive"] }
+hex = { version = "0.4.3", default-features = false }
+lazy_static = { version = "1.1.0", features = ["spin_no_std"] }
+log = { version = "0.4", default-features = false }
+pallet-evm = { default-features = false, git = "https://github.com/integritee-network/frontier.git", branch = "bar/polkadot-v0.9.42" }
+rand = { version = "0.7", optional = true }
+rand-sgx = { package = "rand", git = "https://github.com/mesalock-linux/rand-sgx", tag = "sgx_1.1.3", features = ["sgx_tstd"], optional = true }
+ring = { version = "0.16.20", default-features = false }
+scale-info = { version = "2.4.0", default-features = false, features = ["derive"] }
+secp256k1 = { version = "0.28.0", default-features = false }
+serde = { version = "1.0", default-features = false, features = ["alloc", "derive"] }
+sp-core = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42", default-features = false }
+sp-io = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42", default-features = false }
+sp-runtime = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42", default-features = false }
+sp-std = { git = "https://github.com/paritytech/substrate", branch = "polkadot-v0.9.42", default-features = false }
+strum = { version = "0.26", default-features = false }
+strum_macros = { version = "0.26", default-features = false }
+thiserror = { version = "1.0.26", optional = true }
+
+
+# sgx dependencies
+base64_sgx = { package = "base64", rev = "sgx_1.1.3", git = "https://github.com/mesalock-linux/rust-base64-sgx", optional = true }
+sgx_tstd = { git = "https://github.com/apache/teaclave-sgx-sdk.git", branch = "master", optional = true, features = ["net", "thread"] }
+thiserror-sgx = { package = "thiserror", git = "https://github.com/mesalock-linux/thiserror-sgx", tag = "sgx_1.1.3", optional = true }
+
+
+# internal dependencies
+itp-settings = { path = "../../../core-primitives/settings", default-features = false }
+itp-sgx-crypto = { path = "../../../core-primitives/sgx/crypto", default-features = false }
+itp-sgx-io = { path = "../../../core-primitives/sgx/io", default-features = false }
+itp-utils = { path = "../../../core-primitives/utils", default-features = false }
+litentry-hex-utils = { path = "../../../../primitives/hex", default-features = false }
+parentchain-primitives = { package = "core-primitives", path = "../../../../primitives/core", default-features = false }
+teerex-primitives = { path = "../../../../primitives/teerex", default-features = false }
+# litentry primities 
+litentry-primitives = { path = "../../../litentry/primitives", default-features = false }
+
+[dev-dependencies]
+base64 = { version = "0.13", features = ["alloc"] }
+
+[features]
+default = ["std"]
+production = [
+    "parentchain-primitives/production",
+]
+sgx = [
+    "sgx_tstd",
+    "rand-sgx",
+    "itp-sgx-crypto/sgx",
+    "thiserror-sgx",
+    "itp-sgx-io/sgx",
+    "litentry-primitives/sgx",
+]
+std = [
+    "strum/std",
+    "hex/std",
+    "serde/std",
+    "itp-sgx-crypto/std",
+    "itp-utils/std",
+    "sp-core/std",
+    "sp-std/std",
+    "sp-io/std",
+    "sp-runtime/std",
+    "ring/std",
+    "parentchain-primitives/std",
+    "teerex-primitives/std",
+    "rand",
+    "log/std",
+    "bitcoin/std",
+    "secp256k1/std",
+    "thiserror",
+    "itp-sgx-io/std",
+    "litentry-primitives/std",
+]
